Задание jpa. Учел все замечания.

+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/dao/CommentDao.java
@@ -9,7 +9,7 @@ public interface CommentDao {
 
     Long count();
 
-    Optional<Comment> getById(long id);
+    Optional<Comment> getById(int id);
 
     List<Comment> getAll();
